+++ lib/knapsack_pro/extensions/rspec_extension.rb
@@ -90,7 +90,7 @@ module KnapsackPro
 
           configuration.reporter.report(_expected_example_count = 0) do |reporter|
             configuration.with_suite_hooks do
-              queue_runner.with_batch do |test_file_paths|
+              queue_runner.with_batch do |test_file_paths, queue|
                 knapsack__load_spec_files_batch(test_file_paths)
 
                 examples_passed = ordering_strategy.order(world.example_groups).map do |example_group|
@@ -98,7 +98,12 @@ module KnapsackPro
                   example_group.run(reporter)
                 end.all?
 
-                node_examples_passed = false unless examples_passed
+                if examples_passed
+                  queue.mark_batch_passed
+                else
+                  queue.mark_batch_failed
+                  node_examples_passed = false
+                end
 
                 knapsack__persist_example_statuses