+++ spec/integration/runners/queue/rspec_runner_spec.rb
@@ -739,6 +739,139 @@ describe "#{KnapsackPro::Runners::Queue::RSpecRunner} - Integration tests", :cle
 
       expect(actual.exit_code).to eq 0
     end
+
+    it 'gives access to batch of tests in queue hooks' do
+      rspec_options = ''
+
+      spec_helper = <<~SPEC
+      require 'knapsack_pro'
+      KnapsackPro::Adapters::RSpecAdapter.bind
+
+      KnapsackPro::Hooks::Queue.before_subset_queue do |queue_id, subset_queue_id, queue|
+        print "Tests in batches in before_subset_queue: "
+        puts queue.map(&:test_file_paths).inspect
+
+        print "Batches' statuses in before_subset_queue: "
+        puts queue.map(&:status).inspect
+      end
+
+      KnapsackPro::Hooks::Queue.after_subset_queue do |queue_id, subset_queue_id, queue|
+        print "Tests in batches in after_subset_queue: "
+        puts queue.map(&:test_file_paths).inspect
+        print "Batches' statuses in after_subset_queue: "
+        puts queue.map(&:status).inspect
+
+        # call public API methods that must be backward compatible
+        print "Current batch tests: "
+        puts queue.current_batch.test_file_paths.inspect
+        print "Current batch status: "
+        puts queue.current_batch.status
+      end
+      SPEC
+
+      spec_a = Spec.new('a_spec.rb', <<~SPEC)
+        describe 'A_describe' do
+          it 'A1 test example' do
+            expect(1).to eq 1
+          end
+        end
+      SPEC
+
+      spec_b = Spec.new('b_spec.rb', <<~SPEC)
+        describe 'B_describe' do
+          it 'B1 test example' do
+            expect(1).to eq 1
+          end
+        end
+      SPEC
+
+      spec_c = Spec.new('c_spec.rb', <<~SPEC)
+        describe 'C_describe' do
+          it 'C1 test example' do
+            expect(1).to eq 1
+          end
+        end
+      SPEC
+
+      failing_spec_d = Spec.new('d_spec.rb', <<~SPEC)
+        describe 'D_describe' do
+          it 'D1 test example' do
+            expect(1).to eq 0
+          end
+        end
+      SPEC
+
+      spec_e = Spec.new('e_spec.rb', <<~SPEC)
+        describe 'E_describe' do
+          it 'E1 test example' do
+            expect(1).to eq 1
+          end
+        end
+      SPEC
+
+      spec_f = Spec.new('f_spec.rb', <<~SPEC)
+        describe 'F_describe' do
+          it 'F1 test example' do
+            expect(1).to eq 1
+          end
+        end
+      SPEC
+
+      failing_spec_g = Spec.new('g_spec.rb', <<~SPEC)
+        describe 'G_describe' do
+          it 'G1 test example' do
+            expect(1).to eq 0
+          end
+        end
+      SPEC
+
+      spec_h = Spec.new('h_spec.rb', <<~SPEC)
+        describe 'h_describe' do
+          it 'H1 test example' do
+            expect(1).to eq 1
+          end
+        end
+      SPEC
+
+      generate_specs(spec_helper, rspec_options, [
+        [spec_a, spec_b],
+        [spec_c, failing_spec_d],
+        [spec_e, spec_f],
+        [failing_spec_g, spec_h],
+      ])
+
+      actual = subject
+
+      expect(actual.stdout).to include('Tests in batches in before_subset_queue: [["spec_integration/a_spec.rb", "spec_integration/b_spec.rb"]]')
+      expect(actual.stdout).to include('Tests in batches in before_subset_queue: [["spec_integration/a_spec.rb", "spec_integration/b_spec.rb"], ["spec_integration/c_spec.rb", "spec_integration/d_spec.rb"]]')
+      expect(actual.stdout).to include('Tests in batches in before_subset_queue: [["spec_integration/a_spec.rb", "spec_integration/b_spec.rb"], ["spec_integration/c_spec.rb", "spec_integration/d_spec.rb"], ["spec_integration/e_spec.rb", "spec_integration/f_spec.rb"]]')
+      expect(actual.stdout).to include('Tests in batches in before_subset_queue: [["spec_integration/a_spec.rb", "spec_integration/b_spec.rb"], ["spec_integration/c_spec.rb", "spec_integration/d_spec.rb"], ["spec_integration/e_spec.rb", "spec_integration/f_spec.rb"], ["spec_integration/g_spec.rb", "spec_integration/h_spec.rb"]]')
+
+      expect(actual.stdout).to include('Tests in batches in after_subset_queue: [["spec_integration/a_spec.rb", "spec_integration/b_spec.rb"]]')
+      expect(actual.stdout).to include('Tests in batches in after_subset_queue: [["spec_integration/a_spec.rb", "spec_integration/b_spec.rb"], ["spec_integration/c_spec.rb", "spec_integration/d_spec.rb"]]')
+      expect(actual.stdout).to include('Tests in batches in after_subset_queue: [["spec_integration/a_spec.rb", "spec_integration/b_spec.rb"], ["spec_integration/c_spec.rb", "spec_integration/d_spec.rb"], ["spec_integration/e_spec.rb", "spec_integration/f_spec.rb"]]')
+      expect(actual.stdout).to include('Tests in batches in after_subset_queue: [["spec_integration/a_spec.rb", "spec_integration/b_spec.rb"], ["spec_integration/c_spec.rb", "spec_integration/d_spec.rb"], ["spec_integration/e_spec.rb", "spec_integration/f_spec.rb"], ["spec_integration/g_spec.rb", "spec_integration/h_spec.rb"]]')
+
+
+      expect(actual.stdout).to include("Batches' statuses in before_subset_queue: [:not_executed]")
+      expect(actual.stdout).to include("Batches' statuses in before_subset_queue: [:passed, :not_executed]")
+      expect(actual.stdout).to include("Batches' statuses in before_subset_queue: [:passed, :failed, :not_executed]")
+      expect(actual.stdout).to include("Batches' statuses in before_subset_queue: [:passed, :failed, :passed, :not_executed]")
+
+      expect(actual.stdout).to include("Batches' statuses in after_subset_queue: [:passed]")
+      expect(actual.stdout).to include("Batches' statuses in after_subset_queue: [:passed, :failed]")
+      expect(actual.stdout).to include("Batches' statuses in after_subset_queue: [:passed, :failed, :passed]")
+      expect(actual.stdout).to include("Batches' statuses in after_subset_queue: [:passed, :failed, :passed, :failed]")
+
+      expect(actual.stdout).to include('Current batch tests: ["spec_integration/a_spec.rb", "spec_integration/b_spec.rb"]')
+      expect(actual.stdout).to include('Current batch tests: ["spec_integration/c_spec.rb", "spec_integration/d_spec.rb"]')
+      expect(actual.stdout).to include('Current batch tests: ["spec_integration/e_spec.rb", "spec_integration/f_spec.rb"]')
+      expect(actual.stdout).to include('Current batch tests: ["spec_integration/g_spec.rb", "spec_integration/h_spec.rb"]')
+      expect(actual.stdout).to include('Current batch status: passed').twice
+      expect(actual.stdout).to include('Current batch status: failed').twice
+
+      expect(actual.exit_code).to eq 1
+    end
   end
 
   context 'when the RSpec seed is used' do